// packages/cli/pragma/src/lib/domains/shared/types.ts
export interface StandardSummary {
  readonly uri: URI;
  readonly name: string;
  readonly category: string;
  readonly description: string;
  readonly extends?: string;
}

export interface CodeBlock {
  readonly language: string;
  readonly code: string;
  readonly caption?: string;
}

export interface StandardDetailed extends StandardSummary {
  readonly dos: readonly CodeBlock[];
  readonly donts: readonly CodeBlock[];
}

export interface CategorySummary {
  readonly name: string;
  readonly standardCount: number;
}